+++ src/Control/Monad/ST/Unsafe/Compat.hs
@@ -1,0 +1,9 @@
+{-# LANGUAGE CPP, NoImplicitPrelude #-}
+module Control.Monad.ST.Unsafe.Compat (
+  -- * Unsafe operations
+  unsafeInterleaveST
+, unsafeIOToST
+, unsafeSTToIO
+) where
+
+import Control.Monad.ST (unsafeInterleaveST, unsafeIOToST, unsafeSTToIO)