+++ src/Foreign/ForeignPtr/Unsafe/Compat.hs
@@ -1,0 +1,7 @@
+{-# LANGUAGE CPP, NoImplicitPrelude #-}
+module Foreign.ForeignPtr.Unsafe.Compat (
+  -- ** Unsafe low-level operations
+  unsafeForeignPtrToPtr
+) where
+
+import Foreign.ForeignPtr (unsafeForeignPtrToPtr)