+++ src/Foreign/Marshal/Unsafe/Compat.hs
@@ -1,0 +1,7 @@
+{-# LANGUAGE CPP, NoImplicitPrelude #-}
+module Foreign.Marshal.Unsafe.Compat (
+  -- * Unsafe functions
+  unsafeLocalState
+) where
+
+import Foreign.Marshal (unsafeLocalState)